changing vidoe stream to img
This maybe doing the same thing I have done before in the code above, but it might be a better way of doing it. Changing the video stream to img files to give the machine learning (which will be added later) better pics to pick from. I am also pretty sure I do not have to keep  importing numby,cv2, ect. multiple times. I know there is too much code and all this needs to be cleaned up.